+++ src/Integration/files.jl
@@ -23,11 +23,11 @@ include("BoundaryDescriptors.jl")
 include("BoundaryTriangulations.jl")
 @reexport using Gridap.BoundaryTriangulations
 
-include("NormalVectors.jl")
-@reexport using Gridap.NormalVectors
-
 include("SkeletonTriangulations.jl")
 @reexport using Gridap.SkeletonTriangulations
+
+include("NormalVectors.jl")
+@reexport using Gridap.NormalVectors
 
 include("BoundaryCellFields.jl")
 @reexport using Gridap.BoundaryCellFields