+++ test/MultiFieldTests/MultiFEBasesTests.jl
@@ -104,4 +104,24 @@ ids = ones(Int,ncells(trian))
 cb = CellBasis(trian,σfun,bh[1],ids)
 @test isa(cb,FEBasisWithFieldId)
 
+strian = SkeletonTriangulation(model)
+squad = CellQuadrature(strian,order=2)
+
+sbh = restrict(bh,strian)
+@test isa(sbh[1].cellfield1,FEBasisWithFieldId)
+@test isa(sbh[1].cellfield2,FEBasisWithFieldId)
+
+cm = inner(jump(sbh[1]),jump(sbh[1]))
+ca = integrate(cm,strian,squad)
+
+@test isa(ca.cellmatrix11,MultiCellArray)
+
+uh = zero(V)
+suh = restrict(uh,strian)
+@test isa(suh[1].cellfield1,CellField)
+@test isa(suh[1].cellfield2,CellField)
+
+cm = inner(jump(sbh[1]),jump(suh[1]))
+ca = integrate(cm,strian,squad)
+
 end # module MultiFEBasesTests